Start returning all the data from Speedtest.
This includes all the data from each individual test run.

Previously we were only returning the mean of the median of all the test
runs. This doesn't make much sense.

Now, the highest reporting level can decide how to process the data.

We are grouping unrelated times into a single object as if that
represented one single run, when in reality the "columns" (e.g., "warm")
are all run in a group.