Don't include calls to `random` in the add tests.
Instead pre-cache all the random data we'll need.

Turns out that this was a major performance drag for larger random
sizes, as shown by the blob tests. It was in the top part of the
profile, now its not.

This also ensures that each test-rep within a run uses consistent
random data so compression/hashing/whatever effects should be more
consistent.